Go back to a session per command
The consensus seems to be that requests.Session is not thread safe; at least, not in all circumstances. Although no problems have been experienced with reusing sessions across threads, to be on the safe side, go back to creating a new session for each command. Rename _get_session method to get_session and leave in place in case someone wants to subclass AmcrestCamera and override it.