Add ability to use functionality with NVRs and channels
Allow the Amcrest API to be used with channels.  Mostly, this includes
adding channel parameters to a lot of the API, but only for those
functions that differ per-channel.  In the case of properties with
setters that need to be per-channel, break them out to instead be getter
and setter functions.

Update the interfaces to include type information that can be used for
type checking internally and for external users.

Along with this, do a large-ish update to Python 3-style syntax,
including using f-strings for formatting, adding type annotations to the
interface, and dropping some unnecessary code (passing "utf-8" to
decode, parameters to super, etc).

Also, perform a full `black` formatting.