Use PTZ control command helper
Change the various PTZ functions to use the `ptz_control_command` helper
function rather than directly calling the commands.  Shuffle around some
of the arguments to `ptz_control_command` to make it a bit easier to
use, in particular default all of the `arg`'s to `"0"`, so these only
need to be specified when the value is changed from the default.  This
may break direct uses of `ptz_control_command` for callers using
positional arguments, as the `action` and `code` parameters are moved to
be required, non-default arguments.