Change is_video_enabled to report only main stream
The current is_video_enabled function (and video_enabled property, by
extension) will mix together the main and extra (sub) streams when
determining the channel to use.  As such, the property will report being
disabled when the substream is disabled even if the main stream is
enabled.  Change the behavior of the video_enabled property here to only
report the main stream status on the regular encode.  Add parameters to
the is_video_enabled function to allow the stream and type to be
selected.